Use float64 for RoPE coordinates in half precision

+++ stnet/model/kernels.py
@@ -978,7 +978,7 @@ class MultiScaleRetention(nn.Module):
         )
         half = key_dim // 2
         coord_dtype = (
-            torch.float32 if dtype in (torch.float16, torch.bfloat16) else dtype
+            torch.float64 if dtype in (torch.float16, torch.bfloat16) else dtype
         )
         positions = torch.arange(seq_len, device=device, dtype=coord_dtype)
         inv_freq = 1.0 / self._rope_theta ** torch.linspace(
